example/requirements.txt
SwarmManagement
pyopenssl